finish

+++ src/main/java/pro/sky/java/course2/calculator/CalculatorServiceImpl.java
@@ -24,7 +24,7 @@ public class CalculatorServiceImpl implements CalculatorService {
     }
 
     @Override
-    public int divideNumbers(int num1, int num2) {
-        return num1 / num2;
+    public double divideNumbers(int num1, int num2) {
+        return num1 / (double) num2;
     }
 }
